Refactored folder structure.
