--- WebApp/indoor-localization/src/components/Reports.tsx
@@ -1,14 +1,0 @@
-import Footer from "./Footer";
-
-function Reports() {
-  return (
-    <>
-      <div className="content">
-        <span className="syn-heading--3x-large">Reports</span>
-      </div>
-      <Footer></Footer>
-    </>
-  );
-}
-
-export default Reports;
