+++ WebApp/indoor-localization/src/components/common/Footer.css
