+++ WebApp/indoor-localization/src/components/common/Footer.tsx
