+++ WebApp/indoor-localization/src/components/common/Manager.css
