+++ WebApp/indoor-localization/src/components/common/PopUpConfirmation.tsx
