+++ WebApp/indoor-localization/src/components/reports-page/Reports.tsx
@@ -1,0 +1,14 @@
+import Footer from "../common/Footer";
+
+function Reports() {
+	return (
+		<>
+			<div className="content">
+				<span className="syn-heading--3x-large">Reports</span>
+			</div>
+			<Footer></Footer>
+		</>
+	);
+}
+
+export default Reports;
